feat: add a utility macro for error coercition (#9)
The `try_or_coerce` macro allows to cleanly unwrap a result or convert the error type to one implementing `From` for the current error. 

This is a convenience macro which contributes to writing kernels without having auto-coercion. This is preferable over auto-coercion as this forces user-defined errors to be highlighted compared to regular STM errors